requirements.txt

streamlit
mysql-connector-python
matplotlib
